[client,sdl] rename c++ headers to hpp
