+++ client/SDL/sdl_disp.hpp
@@ -21,7 +21,7 @@
 #include <freerdp/types.h>
 #include <freerdp/client/disp.h>
 
-#include "sdl_freerdp.h"
+#include "sdl_freerdp.hpp"
 
 BOOL sdl_disp_init(sdlDispContext* xfDisp, DispClientContext* disp);
 BOOL sdl_disp_uninit(sdlDispContext* xfDisp, DispClientContext* disp);
