+++ client/SDL/sdl_pointer.cpp
@@ -21,10 +21,10 @@
 
 #include <freerdp/gdi/gdi.h>
 
-#include "sdl_pointer.h"
-#include "sdl_freerdp.h"
-#include "sdl_touch.h"
-#include "sdl_utils.h"
+#include "sdl_pointer.hpp"
+#include "sdl_freerdp.hpp"
+#include "sdl_touch.hpp"
+#include "sdl_utils.hpp"
 
 #include <SDL_mouse.h>
